clean rmsd test
